requirements.txt
opencv-python~=4.7.0.68
numpy~=1.24.1
srt~=3.5.2
alive-progress~=3.0.1
google-cloud-vision~=3.3.1
matplotlib~=3.6.3